+++ src/Models/Deployer/CreateMarketLog.cs
@@ -4,6 +4,7 @@ public struct CreateMarketLog
 {
     [Index] public Address Market;
     [Index] public Address Owner;
+    public Address Router;
     public Address StakingToken;
     public bool AuthPoolCreators;
     public bool AuthProviders;
